Comentários do pattern Visitor e refactor de algumas classes referente a organização do código em pastas.
