+++ src/com/examples/state/Ligado.java
@@ -1,5 +1,7 @@
 package com.examples.state;
 
+import com.examples.state.interfaces.StatusAbstract;
+
 /**
  * Implementa o estado "Ligado". O método executar verifica as regras de negócio antes da mudança
  * do estado.
