--- src/com/examples/visitor/interfaces/Periferico.java
@@ -1,8 +1,0 @@
-package com.examples.visitor.interfaces;
-
-/**
- * Created by ricardo on 2/28/17.
- */
-public interface Periferico  {
-    void accept (PerifericoVisitor visitor);
-}
